pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "relay"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.0.0",
    "typer[all]>=0.9.0,<0.13.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0,<8.2.0",
]

[project.scripts]
relay = "app.main:app"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["documents*", "errors*", "commands*"]
